Pause/Abort functionality added to CPE workflow.
The CPE implementation is terrible and doesn't behave as one would expect given its API so there may still be bugs in it. As examples, removeStatusCallbackListener appear not to function, the paused and resumes callBack methods are never called and if you abort a paused CPE it first resumes, runs for a bit and only then aborts. I'm really hoping UIMA AS is better.